Extract glob logic from rewrites.ts
Also no longer force "/" prefix, since that is specific to paths and prevents are glob logic from being used to match domains, etc ...
